Moving FindRBDMappedFile to ceph/util/
Fixes https://github.com/rook/rook/issues/1056
